# build: optimize .h generation in GenerateHeaderFrom{Raw,Json}.cmake