@l0rinc 2024-09-12
> Seems a bit absurd that the regex replacment is faster than appending to a string

I did some simple profiling and this regex version likely works with some mutable stringbuilder in the background, while the previous string solution was likely quadratic since it's immutable, so it was recreated after each byte.
And the file write is likely faster than that since it uses some buffering.

> also transform the json one

I'm still working on that one, but given the concept ACK, I'll do that here as well.