@l0rinc 2024-09-12
> it would be good to test this on Windows and Linux with the minimum and maximum supported cmake version to make sure this is really the last time this needs to be touched.

While the speed difference is algorithmic, shouldn't be OS dependent, I would appreciate if someone would do that.

> would be more consistent and easier to maintain to also transform the json one

pushed